requirements.txt
discord.py
python-ulid
